+++ frontend/src/pages/GuestReservation/GuestReservation.tsx
@@ -15,7 +15,7 @@ import { GuestMapState } from 'pages/GuestMap/GuestMap';
 import { MapItem, Reservation, ScrollPosition, Space } from 'types/common';
 import { ErrorResponse } from 'types/response';
 import * as Styled from './GuestReservation.styles';
-import ReservationForm from './units/ReservationForm';
+import GuestReservationForm from './units/GuestReservationForm';
 
 interface URLParameter {
   sharingMapId: MapItem['sharingMapId'];
@@ -135,7 +135,7 @@ const GuestReservation = (): JSX.Element => {
           <Styled.ColorDot color={space.color} size="medium" />
           {space.name}
         </Styled.PageHeader>
-        <ReservationForm
+        <GuestReservationForm
           isEditMode={isEditMode}
           space={space}
           reservation={reservation}
